requirements.txt
python-dotenv==0.19.2
langchain==0.2.14
langchain-chroma==0.1.2
langchain-community==0.2.12
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.99
Flask==2.2.5
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-Minify==0.37
flask-restx==1.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1